Removed Win32GameControl.
